keyring: refactor providers
Moves all of the providers previously under `keyring::ed25519` into the
`keyring::providers` module.

This is in anticipation of adding a second signing algorithm (ECDSA).
